Vráceny put na post

+++ routes/web.php
@@ -34,19 +34,26 @@ Route::middleware(['auth:sanctum', 'verified'])->get('/', [DashboardController::
 Route::middleware(['auth:sanctum', 'verified'])->get('/dashboard', [DashboardController::class,'show']) ->name('dashboard');
 
 Route::middleware(['auth:sanctum', 'verified'])->get('/chapters/{id:id}', [ChapterController::class,'showChapters']);
+Route::middleware(['auth:sanctum', 'verified', 'permition:possibility_read'])->get('/chapter/{id:id}', [ChapterController::class,'showChapter']);
+Route::middleware(['auth:sanctum', 'verified', 'permition:possibility_read,edit_content'])->get('/chapter/{id:id}/edit', [ChapterController::class,'showChapterEdit']);
+Route::middleware(['auth:sanctum', 'verified', 'permition:possibility_read'])->get('/chapter/{chapter_id?}/results/{result_id?}', [ChapterController::class,'showChapterResultMe']);
+Route::middleware(['auth:sanctum', 'verified', 'permition:possibility_read'])->get('/chapter/{chapter_id?}/results/{result_id?}/user/{user_id?}', [ChapterController::class,'showChapterResult']);
 
-Route::middleware(['auth:sanctum', 'verified'])->get('/users', [ListUsersController::class,'showAllUsers']);
+Route::middleware(['auth:sanctum', 'verified', 'permition:new_user'])->get('/users', [ListUsersController::class,'showAllUsers']);
 Route::middleware(['auth:sanctum', 'verified', 'permition:new_user'])->get('/users/{id:id}', [ListUsersController::class,'showUser']);
 
 Route::middleware(['auth:sanctum', 'verified', 'permition:edit_permitions'])->get('/permitions', [PermitionController::class,'showPermissions']);
 
 
-Route::middleware(['auth:sanctum', 'verified', 'permition:possibility_read'])->get('/chapter/{id:id}', [ChapterController::class,'showChapter']);
-Route::middleware(['auth:sanctum', 'verified', 'permition:possibility_read,edit_content'])->get('/chapter/{id:id}/edit', [ChapterController::class,'showChapterEdit']);
+//Chat
+Route::middleware(['auth:sanctum', 'verified', 'permition:possibility_read'])->get('/chat/{table_name?}/{element_id?}', [ChatController::class,'getChat']);
+Route::middleware(['auth:sanctum', 'verified', 'permition:possibility_read'])->get('/get_chat_comments/{table_name?}/{element_id?}', [ChatController::class,'getChatComments']);
+Route::middleware(['auth:sanctum', 'verified', 'permition:possibility_read'])->post('/add_comment', [ChatController::class,'addComment']);
+Route::middleware(['auth:sanctum', 'verified', 'permition:possibility_read'])->delete('/remove_comment/{id:id}', [ChatController::class,'removeComment']);
 
-
-Route::middleware(['auth:sanctum', 'verified', 'permition:possibility_read'])->get('/chapter/{chapter_id?}/results/{result_id?}', [ChapterController::class,'showChapterResultMe']);
-Route::middleware(['auth:sanctum', 'verified', 'permition:possibility_read'])->get('/chapter/{chapter_id?}/results/{result_id?}/user/{user_id?}', [ChapterController::class,'showChapterResult']);
+//Výsledky
+Route::middleware(['auth:sanctum', 'verified', 'permition:possibility_read'])->post('/add_result', [ResultsController::class,'addResult']);
+Route::middleware(['auth:sanctum', 'verified', 'permition:possibility_read'])->post('/add_test_result', [ResultsController::class,'addTestResult']);
 
 Route::middleware(['auth:sanctum', 'verified', 'permition:possibility_read'])->get('/chapter/{id:id}/all_results', [ResultsController::class,'showAllResults']);
 Route::middleware(['auth:sanctum', 'verified', 'permition:possibility_read'])->get('/chapter/{id:id}/all_results/{element_id?}', [ResultsController::class,'showResults']);
@@ -54,55 +61,44 @@ Route::middleware(['auth:sanctum', 'verified', 'permition:edit_content'])->get('
 Route::middleware(['auth:sanctum', 'verified', 'permition:edit_content'])->get('/element_abc_results/{element_id?}', [ResultsController::class,'showABCResults']);
 
 //Kontent
+Route::middleware(['auth:sanctum', 'verified', 'permition:edit_content'])->get('/edit_setting/{table_name?}/{id:id}', [ContentController::class,'editSetting']);
+Route::middleware(['auth:sanctum', 'verified', 'permition:edit_content'])->post('/save_setting/{table_name?}/{id:id}', [ContentController::class,'saveSetting']);
+Route::middleware(['auth:sanctum', 'verified', 'permition:edit_content'])->post('/move', [ContentController::class,'move']);
 
-Route::middleware(['auth:sanctum', 'verified', 'permition:possibility_read'])->get('/chat/{table_name?}/{element_id?}', [ChatController::class,'getChat']);
-Route::middleware(['auth:sanctum', 'verified', 'permition:possibility_read'])->get('/get_chat_comments/{table_name?}/{element_id?}', [ChatController::class,'getChatComments']);
-Route::middleware(['auth:sanctum', 'verified', 'permition:possibility_read'])->post('/add_comment', [ChatController::class,'addComment']);
-Route::middleware(['auth:sanctum', 'verified', 'permition:possibility_read'])->delete('/remove_comment/{id:id}', [ChatController::class,'removeComment']);
+Route::middleware(['auth:sanctum', 'verified', 'permition:edit_content'])->post('/add_element', [ContentController::class,'addElement']);
+Route::middleware(['auth:sanctum', 'verified', 'permition:edit_content'])->post('/get_element_selector', [ContentController::class,'getElementsSelector']);
+
+Route::middleware(['auth:sanctum', 'verified', 'permition:edit_content'])->delete('/remove/{table_name?}/{id:id}', [ContentController::class,'removeElement']);
+Route::middleware(['auth:sanctum', 'verified', 'permition:edit_content'])->delete('/remove_image/{id:id}', [ContentController::class,'removeImage']);
+Route::middleware(['auth:sanctum', 'verified', 'permition:edit_content'])->delete('/remove_file/{id:id}', [ContentController::class,'removeFile']);
 
 Route::middleware(['auth:sanctum', 'verified', 'permition:edit_content'])->post('/save_image', [ContentController::class,'saveImage']);
 Route::middleware(['auth:sanctum', 'verified', 'permition:edit_content'])->post('/add_image', [ContentController::class,'addImage']);
 Route::middleware(['auth:sanctum', 'verified', 'permition:edit_content'])->post('/save_name', [ContentController::class,'saveName']);
 Route::middleware(['auth:sanctum', 'verified', 'permition:edit_content'])->post('/save_description', [ContentController::class,'saveDescription']);
 Route::middleware(['auth:sanctum', 'verified', 'permition:edit_content'])->post('/save_column', [ContentController::class,'saveColumn']);
-Route::middleware(['auth:sanctum', 'verified', 'permition:possibility_read'])->post('/save_finished', [ContentController::class,'saveFinished']);
-Route::middleware(['auth:sanctum', 'verified', 'permition:possibility_read'])->post('/add_result', [ResultsController::class,'addResult']);
-Route::middleware(['auth:sanctum', 'verified', 'permition:possibility_read'])->post('/add_test_result', [ResultsController::class,'addTestResult']);
-
 
 Route::middleware(['auth:sanctum', 'verified', 'permition:possibility_read' ])->post('/save_file', [ContentController::class,'saveFile']);
 Route::middleware(['auth:sanctum', 'verified', 'permition:possibility_read' ])->post('/save_file_result', [ContentController::class,'saveFileResult']);
 Route::middleware(['auth:sanctum', 'verified', 'permition:possibility_read' ])->post('/add_file', [ContentController::class,'addFile']);
 
-Route::middleware(['auth:sanctum', 'verified', 'permition:edit_content'])->delete('/remove/{table_name?}/{id:id}', [ContentController::class,'removeElement']);
-Route::middleware(['auth:sanctum', 'verified', 'permition:edit_content'])->delete('/remove_image/{id:id}', [ContentController::class,'removeImage']);
-Route::middleware(['auth:sanctum', 'verified', 'permition:edit_content'])->delete('/remove_file/{id:id}', [ContentController::class,'removeFile']);
-
-Route::middleware(['auth:sanctum', 'verified', ])->post('/finish_element', [ContentController::class,'finishElement']);
-
-Route::middleware(['auth:sanctum', 'verified'])->get('/image_selector', [ContentController::class,'getImagesSelector']);
-Route::middleware(['auth:sanctum', 'verified'])->get('/image_selector_gallery', [ContentController::class,'getImagesSelectorGallery']);
-
-Route::middleware(['auth:sanctum', 'verified'])->get('/file_selector', [ContentController::class,'getFileSelector']);
-Route::middleware(['auth:sanctum', 'verified'])->get('/file_selector_gallery', [ContentController::class,'getFileSelectorGallery']);
-
-Route::middleware(['auth:sanctum', 'verified', 'permition:edit_content'])->post('/add_element', [ContentController::class,'addElement']);
-
-Route::middleware(['auth:sanctum', 'verified', 'permition:edit_content'])->post('/get_element_selector', [ContentController::class,'getElementsSelector']);
+Route::middleware(['auth:sanctum', 'verified', 'permition:possibility_read'])->post('/save_finished', [ContentController::class,'saveFinished']);
+Route::middleware(['auth:sanctum', 'verified', 'permition:possibility_read' ])->post('/finish_element', [ContentController::class,'finishElement']);
 
 
-Route::middleware(['auth:sanctum', 'verified', 'permition:edit_content'])->post('/move', [ContentController::class,'move']);
+Route::middleware(['auth:sanctum', 'verified', 'permition:possibility_read'])->get('/image_selector', [ContentController::class,'getImagesSelector']);
+Route::middleware(['auth:sanctum', 'verified', 'permition:possibility_read'])->get('/image_selector_gallery', [ContentController::class,'getImagesSelectorGallery']);
 
-Route::middleware(['auth:sanctum', 'verified', 'permition:edit_content'])->get('/edit_setting/{table_name?}/{id:id}', [ContentController::class,'editSetting']);
-Route::middleware(['auth:sanctum', 'verified', 'permition:edit_content'])->post('/save_setting/{table_name?}/{id:id}', [ContentController::class,'saveSetting']);
+Route::middleware(['auth:sanctum', 'verified', 'permition:possibility_read'])->get('/file_selector', [ContentController::class,'getFileSelector']);
+Route::middleware(['auth:sanctum', 'verified', 'permition:possibility_read'])->get('/file_selector_gallery', [ContentController::class,'getFileSelectorGallery']);
 
 //Učebnice
 Route::middleware(['auth:sanctum', 'verified', 'permition:edit_content'])->post('/add_book', [BooksController::class,'addBook']);
-Route::middleware(['auth:sanctum', 'verified'])->get('/get_book_status/{id:id}', [BooksController::class,'getStatus']);
+Route::middleware(['auth:sanctum', 'verified', 'permition:edit_content'])->get('/get_book_status/{id:id}', [BooksController::class,'getStatus']);
 
 //Kapitoly
 Route::middleware(['auth:sanctum', 'verified', 'permition:edit_content'])->post('/add_chapter', [ChapterController::class,'addChapter']);
-Route::middleware(['auth:sanctum', 'verified'])->get('/get_chapter_status/{id:id}', [ChapterController::class,'getStatus']);
+Route::middleware(['auth:sanctum', 'verified', 'permition:edit_content'])->get('/get_chapter_status/{id:id}', [ChapterController::class,'getStatus']);
 
 
 //Pravidla
@@ -113,14 +109,14 @@ Route::middleware(['auth:sanctum', 'verified'])->post('/save_rule', [LockControl
 Route::middleware(['auth:sanctum', 'verified'])->post('/unlock', [LockController::class,'unlock']);
 
 //Uživatelé
-Route::post('/users/{id:id}/saveUserData', [ListUsersController::class,'saveUserData']);
-Route::get('/users/usersSort/{sort?}', [ListUsersController::class,'usersSort']);
-Route::get('/users/usersFind/{find?}', [ListUsersController::class,'usersFind']);
-Route::get('/getUserNames', [ListUsersController::class,'getUserNames']);
-Route::middleware(['auth:sanctum', 'verified'])->get('/get_user_status/{id:id}', [ListUsersController::class,'getStatus']);
+Route::middleware(['auth:sanctum', 'verified', 'permition:new_user'])->post('/users/{id:id}/saveUserData', [ListUsersController::class,'saveUserData']);
+Route::middleware(['auth:sanctum', 'verified', 'permition:new_user'])->get('/users/usersSort/{sort?}', [ListUsersController::class,'usersSort']);
+Route::middleware(['auth:sanctum', 'verified', 'permition:new_user'])->get('/users/usersFind/{find?}', [ListUsersController::class,'usersFind']);
+Route::middleware(['auth:sanctum', 'verified', 'permition:new_user'])->get('/getUserNames', [ListUsersController::class,'getUserNames']);
+Route::middleware(['auth:sanctum', 'verified', 'permition:new_user'])->get('/get_user_status/{id:id}', [ListUsersController::class,'getStatus']);
 
 
 //Oprávnění
-Route::post('/addPermition', [PermitionController::class,'addPermition']);
-Route::post('/savePermitionData', [PermitionController::class,'savePermitionData']);
-Route::delete('/removePermition/{id:id}', [PermitionController::class,'removePermition']);
+Route::middleware(['auth:sanctum', 'verified', 'permition:edit_permitions'])->post('/addPermition', [PermitionController::class,'addPermition']);
+Route::middleware(['auth:sanctum', 'verified', 'permition:edit_permitions'])->post('/savePermitionData', [PermitionController::class,'savePermitionData']);
+Route::middleware(['auth:sanctum', 'verified', 'permition:edit_permitions'])->delete('/removePermition/{id:id}', [PermitionController::class,'removePermition']);
